refactor: rename regression test files and update namespaces to Refitter.Tests.Regression
