+++ src/Refitter.Tests/Regression/OneOfDiscriminatorNullRefTests.cs
@@ -2,9 +2,8 @@ using FluentAssertions;
 using Refitter.Core;
 using Refitter.Tests.Build;
 using Refitter.Tests.TestUtilities;
-using TUnit.Core;
 
-namespace Refitter.Tests.RegressionTests;
+namespace Refitter.Tests.Regression;
 
 /// <summary>
 /// Regression tests for Issue #1015: NullReferenceException on Swagger 2.0 documents
