+++ src/Refitter.Tests/Regression/OpenApiTitleAngleBracketIdentifierTests.cs
@@ -2,9 +2,8 @@ using FluentAssertions;
 using Refitter.Core;
 using Refitter.Tests.Build;
 using Refitter.Tests.TestUtilities;
-using TUnit.Core;
 
-namespace Refitter.Tests.RegressionTests;
+namespace Refitter.Tests.Regression;
 
 public class OpenApiTitleAngleBracketIdentifierTests
 {
